requirements.txt
discord.py
Pillow
piexif
rawpy
exifread
